fix splitting the last column in a table when the previous column is an auto-deleted empty column
The initial value for the index of the new column (to the right of the
new boundary) was being set incorrectly in this case because I had used
numCols to guard a call to colBoundXs.get(), but should have used
colBoundXs.size(), which is usually one larger than numCols because it
includes the column boundary before the whole table. So newColIndex was
one less than it should have been, and newColIndex-1 indexed the column
two before the new boundary instead of one before, which can be an
auto-deleted empty column. That caused the SplitColumn edit to fail
because it looked like it was splitting nothing off the column on the
left side of the new boundary.